Fix/25 - add unit tests for  make_dataset.py (#29)
* First unit test for make_dataset.py

* Unit test str_to_dict

* Closes issue #26

* Add a small dataset to use for testing
Prior to this change, we lacked unit tests for some functions in the
make_dataset.py. These were the pipeline parts that required a
"dataframe" as input. Column names were hard coded.

This change adds a pickled pandas dataframe that can be read in using pandas.
This data originates from a query run by FL, file named
test_page_meta_htype_2018-10-15.csv.gz . The file was in using pd.read_csv('test_page_meta_htype_2018-10-15.csv.gz', nrows=100, compression='gzip',
                   error_bad_lines=False). And then pickled with
                   df.to_pickle("user_journey_df.pkl"). At 100 rows it
                   should be small enough for the unit tests to run
                   quickly.

* Pandas test df is read-in

* Add failing sequence_preprocess test

* Use sha256 of pd df to test changes

* Update readme to describe pkl origins

* Add notebook for reproducibility but ignored

* Apply linter to test scripts

* Comment out test_sequence_preprocess

test_sequence_preprocess is commented out of test_make_dataset.py as the
test is failing as NameError: name 'logger' is not defined, this may be
due logging structure or how the tests scope for variables.

We could consider defining a logger for each function, and above that
for each class in the init (if we use classes later, or even for
    functions with scripts).

    So if we define a logger at the top of the script with
    logging.getLogger("make_dataset") you can define another one lower
    down with logging.getLogger("make_dataset.function"). This will work
    as logging is hierarchical.

    We could use the same approach for the unit tests.

    Could also consider an env variable to set the logging level.